mobile/statistics: show statistics full screen
Because of the way Kirigami works with the PageStack, this can only
happen for the first page. So we need to make the statistics page the
only page if selected. And similarly clear it off the stack when
switching to a different page.

There is some awkward special case handling to ensure that the DiveList
is otherwise the first page.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>