mobile/statistics: create an adaptive layout depending on screen orientation
This adds labels for the combo boxes and converts the whole layout into
something a bit more attractive that switches the screen positions of
the various elements depending on screen orientation.

It also does away with the plot button since all changes to the combo
boxes already trigger a plot, so all we need to do is trigger a first
one when the page becomes visible.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>