mobile/UI: add helper function to refresh dive details
This seems overly complex, but it basically makes sure that if we have a
DiveDetails page shown, it is temporarily moved to the end of the page
stack, then replaced with a new instance (which as a side effect redraws
the profile), and then moved back to where it was. This happens without
creating visual artifacts in my testing.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>